Fix unneeded libluaconfig.o presense in libing-gen-utils.so linking
CURDIR variable with absolute path to src/c folder is used on add
libluaconfig.c into SRC_LUACONFIG list.

Because of absolute path it cannot be filtered and "libluaconfig.o"
present in libing-gen-utils.so.
That cause linking error for packages depends on libing-gen-utils
but not linked libconfig explicitly

Remove CURDIR variable usage

Signed-off-by: Ivan Efimov <i.efimov@inango-systems.com>